mcp_server/pyproject.toml
[project]
name = "mcp_server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "fastmcp>=2.10.0",
    "openai>=2.1.0",
]
